Make manifest_parser_perftest respect NINJA_PYTHON
The NINJA_PYTHON define is set by passing --with-python to configure.py.
The default value is set to the python executable running configure.py.

The problem is that one perf test was hardcoding "python" instead of
using NINJA_PYTHON.

To reproduce, run with and without this patch the following commands:

  git clean -f -d -x
  python3 configure.py --bootstrap
  ./ninja manifest_parser_perftest
  ./manifest_parser_perftest

Without this change, output is:

  Creating manifest data...sh: 1: python: not found